Add progress bar to initial loading
