+++ frontend/src/App.js
@@ -3,6 +3,7 @@
 import React, { Component, Fragment } from 'react';
 import './App.css';
 import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom';
+import { LinearProgress } from '@material-ui/core';
 
 
 import AuthPage from './pages/auth/Auth';
@@ -11,12 +12,42 @@ import ListsPage from './pages/lists/Lists';
 
 import MainNavigation from './components/nav/MainNav';
 import AuthContext from './context/auth-context';
+import { fetchRequest } from './utils/helper';
 
 class App extends Component {
 
 	state = {
-		userID: ''
+		userID: '',
+		isLoading: true
 	};
+
+	componentDidMount() {
+		const requestBody = {
+			query: `
+				query {
+					resumeID
+				}
+			`
+		};
+		fetchRequest(requestBody).
+		then((res) => {
+			if (res.status !== 200 && res.status !== 201) {
+				throw new Error('Failed');
+			}
+			return res.json();
+		}).
+		then((resbody) => {
+			const resdata = resbody.data;
+			if (resdata.resumeID) {
+				this.authenticate(resdata.resumeID);
+			}
+		}).then(() => {
+			this.setState({ isLoading: false });
+		}).
+		catch((err) => {
+			console.log(err);
+		});
+	}
 
 	authenticate = (userID) => {
 		this.setState({
@@ -27,13 +58,13 @@ class App extends Component {
 	signout = () => {
 		this.setState({
 			userID: ''
-		})
+		});
 	};
 
   render() {
     return (
-			<BrowserRouter>
-				<Fragment>
+		<BrowserRouter>
+			<Fragment>
 					<AuthContext.Provider value={{
 						userID: this.state.userID,
 						authenticate: this.authenticate,
@@ -41,24 +72,27 @@ class App extends Component {
 					}}>
 						<Switch>
 							<main className="main-content">
-								{!this.state.userID && <Redirect to="/auth" exact />}
-								{!this.state.userID && <Route path="/auth" component={AuthPage} />}
-								{this.state.userID && <Redirect path="/" to="/home" exact />}
-								{this.state.userID && <Redirect path="/auth" to="/home" exact />}
-								{this.state.userID &&
-									<Fragment>
-										<Route path="/home" component={HomePage} />
-										<Route path="/lists" component={ListsPage} />
-										<Route path="/search" component={null} />
-										<Route path="/profile" component={null} />
-									</Fragment>
-								}
+								{this.state.isLoading && <LinearProgress /> }
+								{!this.state.isLoading && <Fragment>
+									{!this.state.userID && <Redirect to="/auth" exact />}
+									{!this.state.userID && <Route path="/auth" component={AuthPage} />}
+									{this.state.userID && <Redirect path="/" to="/home" exact />}
+									{this.state.userID && <Redirect path="/auth" to="/home" exact />}
+									{this.state.userID &&
+										<Fragment>
+											<Route path="/home" component={HomePage} />
+											<Route path="/lists" component={ListsPage} />
+											<Route path="/search" component={null} />
+											<Route path="/profile" component={null} />
+										</Fragment>
+									}
+								</Fragment>}
 							</main>
 						</Switch>
 						<MainNavigation />
-				</AuthContext.Provider>
-				</Fragment>
-			</BrowserRouter>
+					</AuthContext.Provider>
+			</Fragment>
+		</BrowserRouter>
     );
   }
